feat(api): change type of some props

+++ API/Entities/ProductNutrition.cs
@@ -6,9 +6,9 @@ public class ProductNutrition
     public int ProductNutritionCategoryId { get; set; }
     public string Name { get; set; }
     public int CaloriesPer100Grams { get; set; }
-    public int ProteinPer100Grams { get; set; }
-    public int FatPer100Grams { get; set; }
-    public int CarbohydratesPer100Grams { get; set; }
+    public double ProteinPer100Grams { get; set; }
+    public double FatPer100Grams { get; set; }
+    public double CarbohydratesPer100Grams { get; set; }
 
     public ProductNutritionCategory ProductNutritionCategory { get; set; }
